+++ scripts/deploy_localnet.sh
@@ -189,3 +189,4 @@ else
     exit 1
 fi
 
+